Added cpp sim folder
