+++ sim/python/demo_ARCLAB_cone.py
@@ -60,7 +60,7 @@ class ChCameraSensor_DataGeneratorFunctor(veh.ChExternalDriver_DataGeneratorFunc
         rgba8_buffer = self.cam.GetMostRecentRGBA8Buffer()
         if rgba8_buffer.HasData():
             rgba8_data = rgba8_buffer.GetRGBA8Data()
-            rgba8_data = np.ascontiguousarray(np.flipud(rgba8_data)) # Needs to be contigious
+            rgba8_data = np.ascontiguousarray(rgba8_data[::-1]) # Needs to be contigious
             shape = rgba8_data.shape
             writer.Key("width") << shape[1]
             writer.Key("height") << shape[0] 
@@ -166,8 +166,6 @@ class ChDriverInputs_DataParserFunctor(veh.ChExternalDriver_DataParserFunctor):
 
         self.driver = driver
 
-
-
     def Deserialize(self, reader):
         steering = reader.GetFloat()
         throttle = reader.GetFloat()
@@ -184,8 +182,8 @@ def main():
     
     def AddRandomCones(count, filename, class_id=1):
         mmesh = chrono.ChTriangleMeshConnected()
-        mmesh.SetLoadWavefrontMesh(filename, False, True)  
-        mmesh.SetTransform(chrono.ChVectorD(0, 0, 0), chrono.ChMatrix33(1)) 
+        mmesh.LoadWavefrontMesh(filename, False, True)  
+        mmesh.Transform(chrono.ChVectorD(0, 0, 0), chrono.ChMatrix33D(1)) 
         for i in range(count):
             # Generate a random position
             x = (chrono.ChRandom() - .5) * cone_spread_x + cone_offset_x
@@ -196,22 +194,22 @@ def main():
             trimesh_shape = chrono.ChTriangleMeshShape()
             trimesh_shape.SetMesh(mmesh)
             trimesh_shape.SetName(filename)
-            trimesh_shape.SetStatic(True)
+            trimesh_shape.SetMutable(False)
             trimesh_shape.SetScale(chrono.ChVectorD(1, 1, 1))
 
             mesh_body = chrono.ChBody()
             mesh_body.SetPos(pos)
-            mesh_body.SetRot(chrono.ChQuaternion(1, 0, 0, 0))
-            mesh_body.AddAsset(trimesh_shape)
+            mesh_body.SetRot(chrono.ChQuaternionD(1, 0, 0, 0))
+            mesh_body.AddVisualShape(trimesh_shape)
             mesh_body.SetBodyFixed(True)
             vehicle.GetSystem().Add(mesh_body)
 
             if (class_id == 1):
                 red_cone_assets.append(trimesh_shape)
-                cones.append(mesh_body)
+                red_cones.append(mesh_body)
             else:
                 green_cone_assets.append(trimesh_shape)
-                cones.append(mesh_body)
+                green_cones.append(mesh_body)
 
     def AddConesFromFile():
         green_cone_mesh = chrono.ChTriangleMeshConnected()
@@ -222,44 +220,60 @@ def main():
         red_cone_mesh.LoadWavefrontMesh(chrono.GetChronoDataFile("sensor/cones/red_cone.obj"), False, True)
         red_cone_mesh.Transform(chrono.ChVectorD(0, 0, 0), chrono.ChMatrix33D(1))
 
-        with open(chrono.GetChronoDataFile(cone_path_file)) as cone_file:
+        with open(chrono.GetChronoDataFile("autonomy-toolkit/paths/cone_paths_0.csv")) as cone_file:
             while True:
                 line = cone_file.readline()
                 if not line:
                     break
-                id, color, x, y = map(float, line.split('\t')) 
+                x_inner, y_inner, x_outer, y_outer = map(float, line.split(' ')) 
 
-                pos_x = x + cone_offset_x
-                pos_y = y + cone_offset_y
-                pos_z = terrain.GetHeight(chrono.ChVectorD(pos_x, pos_y, 1000))
+                pos_green_x = x_inner + cone_offset_x
+                pos_green_y = y_inner + cone_offset_y
+                pos_green_z = terrain.GetHeight(chrono.ChVectorD(pos_green_x, pos_green_y, 1000))
 
-                pos = chrono.ChVectorD(pos_x, pos_y, pos_z)
+                pos_red_x = x_outer + cone_offset_x
+                pos_red_y = y_outer + cone_offset_y
+                pos_red_z = terrain.GetHeight(chrono.ChVectorD(pos_red_x, pos_red_y, 1000))
+
+                pos_green = chrono.ChVectorD(pos_green_x, pos_green_y, pos_green_z)
+                pos_red = chrono.ChVectorD(pos_red_x, pos_red_y, pos_red_z)
                 rot = chrono.ChQuaternionD(1, 0, 0, 0)
 
-                cone_shape = chrono.ChTriangleMeshShape()
-                if (color == 0):
-                    cone_shape.SetMesh(red_cone_mesh)
-                    red_cone_assets.append(cone_shape)
-                else:
-                    cone_shape.SetMesh(green_cone_mesh)
-                    green_cone_assets.append(cone_shape)
+                green_cone_shape = chrono.ChTriangleMeshShape()
+                green_cone_shape.SetMesh(green_cone_mesh)
+                green_cone_shape.SetName("green_cone_shape")
+                # green_cone_shape.SetStatic(True)
+                green_cone_shape.SetMutable(False)
 
-                cone_shape.SetStatic(True)
+                green_cone_assets.append(green_cone_shape)
+                green_body = chrono.ChBody()
+                green_body.SetPos(pos_green)
+                green_body.SetRot(rot)
+                green_body.AddVisualShape(green_cone_shape)
+                green_body.SetBodyFixed(True)
+                green_cones.append(green_body)
+                vehicle.GetSystem().Add(green_body)
 
-                cone_body = chrono.ChBody()
-                cone_body.SetPos(pos)
-                cone_body.SetRot(rot)
-                cone_body.AddAsset(cone_shape)
-                cone_body.SetBodyFixed(True)
-                cones.append(cone_body)
-                vehicle.GetSystem().Add(cone_body)
+                red_cone_shape = chrono.ChTriangleMeshShape()
+                red_cone_shape.SetMesh(red_cone_mesh)
+                red_cone_shape.SetName("red_cone_shape")
+                # red_cone_shape.SetStatic(True)
+                green_cone_shape.SetMutable(False)
 
+                red_cone_assets.append(red_cone_shape)
+                red_body = chrono.ChBody()
+                red_body.SetPos(pos_red)
+                red_body.SetRot(rot)
+                red_body.AddVisualShape(red_cone_shape)
+                red_body.SetBodyFixed(True)
+                red_cones.append(red_body)
+                vehicle.GetSystem().Add(red_body)
             
     def LabelConeAssets():
             cone_id = 0
             for cone in red_cone_assets:
                 cone_id += 1
-                for mat in cone.material_list:
+                for mat in cone.GetMaterials():
                     mat.SetClassID(1)
                     mat.SetInstanceID(cone_id)
 
@@ -267,9 +281,23 @@ def main():
             cone_id = 0
             for cone in green_cone_assets:
                 cone_id += 1
-                for mat in cone.material_list:
+                for mat in cone.GetMaterials():
                     mat.SetClassID(2)
                     mat.SetInstanceID(cone_id)
+
+    def RedistributeCones():
+        for cone in green_cones:
+            x = (chrono.ChRandom() - .5) * cone_spread_x + cone_offset_x
+            y = (chrono.ChRandom() - .5) * cone_spread_y + cone_offset_y
+            z = terrain.GetHeight((x, y, 1000))
+            cone.SetPos((x, y, z))
+
+        for cone in red_cones:
+            x = (chrono.ChRandom() - .5) * cone_spread_x + cone_offset_x
+            y = (chrono.ChRandom() - .5) * cone_spread_y + cone_offset_y
+            z = terrain.GetHeight((x, y, 1000))
+            cone.SetPos((x, y, z))
+
 
 
     # Create the RCCar vehicle, set parameters, and initialize
@@ -297,47 +325,38 @@ def main():
     patch_mat.SetRestitution(0.01) #-----check if mu, cr, and y is set similar to minfo
 
     # not right need to introduce point cloud -- not sure how to access this using the package_share_directory
-    patch = terrain.AddPatch(patch_mat,
-                             chrono.ChVectorD(
-                                 0, 0, 0), chrono.ChVectorD(0, 0, 1),
-                             terrainLength, terrainWidth)
-    patch.SetTexture(veh.GetDataFile("autonomy-toolkit/terrain/textures/tile4.jpg"), 200, 200)
-
-
-    patch.SetColor(chrono.ChColor(0.8, 0.8, 0.5))
-    #patch = terrain.AddPatch(patch_mat, chrono.CSYSNORM, chrono.GetChronoDataFile("autonomy-toolkit/me3038/me3038_stereo_10cm_2.obj"))  #chrono.ChCoordsysD(chrono.ChVectorD(0, 0, 0), chrono.ChQuaternionD(1, 0, 0, 0))
-    # patch = terrain.AddPatch(patch_mat, chrono.CSYSNORM, chrono.GetChronoDataFile("autonomy-toolkit/me3038/rm3038_pt_cloud_2.obj"))
+    patch = terrain.AddPatch(patch_mat, chrono.CSYSNORM, chrono.GetChronoDataFile("autonomy-toolkit/me3038/rm3038_pt_cloud.obj"))  #chrono.ChCoordsysD(chrono.ChVectorD(0, 0, 0), chrono.ChQuaternionD(1, 0, 0, 0))
     # patch.SetTexture(veh.GetDataFile("terrain/textures/tile4.jpg"), 200, 200)
     # patch.SetColor(chrono.ChColor(0.8, 0.8, 0.5))
     terrain.Initialize()
 
     # add in ME3038 room mesh
-    #room_mmesh = chrono.ChTriangleMeshConnected()
-    #room_mmesh.LoadWavefrontMesh(chrono.GetChronoDataFile("autonomy-toolkit/me3038/rm3038_pt_cloud.obj"), False, True)
-    #room_mmesh.Transform(chrono.ChVectorD(0, 0, 0), chrono.ChMatrix33D(1))
+    room_mmesh = chrono.ChTriangleMeshConnected()
+    room_mmesh.LoadWavefrontMesh(chrono.GetChronoDataFile("autonomy-toolkit/me3038/rm3038_pt_cloud.obj"), False, True)
+    room_mmesh.Transform(chrono.ChVectorD(0, 0, 0), chrono.ChMatrix33D(1))
 
-    #room_trimesh_shape = chrono.ChTriangleMeshShape()
-    #room_trimesh_shape.SetMesh(room_mmesh)
-    #room_trimesh_shape.SetName("ME3038")
-    #room_trimesh_shape.SetStatic(True)
-    #room_mesh_body = chrono.ChBody()
-    #room_mesh_body.SetPos(chrono.ChVectorD(0, 0, 0))
-    #room_mesh_body.AddAsset(room_trimesh_shape)
-    #room_mesh_body.SetBodyFixed(True)
+    room_trimesh_shape = chrono.ChTriangleMeshShape()
+    room_trimesh_shape.SetMesh(room_mmesh)
+    room_trimesh_shape.SetName("ME3038")
+    room_trimesh_shape.SetMutable(False)
 
-    #vehicle.GetSystem().Add(room_mesh_body)
+    room_mesh_body = chrono.ChBody()
+    room_mesh_body.SetPos(chrono.ChVectorD(0, 0, 0))
+    room_mesh_body.AddVisualShape(room_trimesh_shape)
+    room_mesh_body.SetBodyFixed(True)
+
+    vehicle.GetSystem().Add(room_mesh_body)
 
 
     if cones_from_file:
         AddConesFromFile()
-    #else:
-    #    AddRandomCones(num_cones / 2, chrono.GetChronoDataFile("sensor/cones/green_cone.obj"), 2)
-    #    AddRandomCones(num_cones / 2, chrono.GetChronoDataFile("sensor/cones/red_cone.obj"), 1)
-
+    else:
+        AddRandomCones(int(num_cones / 2), chrono.GetChronoDataFile("sensor/cones/green_cone.obj"), 2)
+        AddRandomCones(int(num_cones / 2), chrono.GetChronoDataFile("sensor/cones/red_cone.obj"), 1)
 
     # === create sensors ===
     manager = sens.ChSensorManager(vehicle.GetSystem())
-    manager.scene.AddPointLight(chrono.ChVectorF(0, 0, 100), chrono.ChVectorF(1, 1, 1), 5000)
+    manager.scene.AddPointLight(chrono.ChVectorF(100, 100, 100), chrono.ChColor(1, 1, 1), 5000)
     
     b = sens.Background()
     b.color_horizon = chrono.ChVectorF(.6, .7, .8)
@@ -345,12 +364,13 @@ def main():
     b.mode = sens.BackgroundMode_GRADIENT
     manager.scene.SetBackground(b)
 
-    camera_pose = chrono.ChFrameD(chrono.ChVectorD(0.204, 0, 0.10018), chrono.Q_from_AngAxis(.1, chrono.ChVectorD(0, 1, 0)))
+    camera_pose = chrono.ChFrameD(chrono.ChVectorD(0.204, 0, 0.10018), chrono.Q_from_AngAxis(.2, chrono.ChVectorD(0, 1, 0)))
     width = 1280
     height = 720
     frame_rate = 30.0
-    fov = 1.396  # 80 degree FOV camera)
-    # fov = 1.59
+    if create_semantic_maps:
+        frame_rate = float(5)    # not sure about this
+    fov = 1.396;  # 80 degree FOV camera)
 
     camera = sens.ChCameraSensor(
                 vehicle.GetChassisBody(), # body camera is attached to
@@ -359,27 +379,13 @@ def main():
                 width,                      # image width
                 height,                     # image height
                 fov,                        # FOV
-                2,
-                sens.RADIAL)  # super sample diameter
-    camera.SetRadialLensParameters(chrono.ChVectorF(-0.369, 0.1257, -0.0194))
+                2)
+
     camera.SetName("Camera Sensor")
     c_window = float(0)
     camera.SetCollectionWindow(c_window)
-    # camera.SetLag(.5)
 
-    lidar = sens.ChLidarSensor(
-                vehicle.GetChassisBody(), # body camera is attached to
-                frame_rate,                 # update rate in Hz
-                camera_pose,                # offset pose
-                width,                      # image width
-                height,                     # image height
-                fov,                        # FOV
-                2,
-                sens.RADIAL
-    )
-
-    if vis:
-        camera.PushFilter(sens.ChFilterVisualize(640, 360))
+    # camera.PushFilter(sens.ChFilterVisualize(640, 360))
 
     if save_sensor_data:
         camera.PushFilter(sens.ChFilterSave(sensor_data_dir + "cam1/"))
@@ -391,23 +397,20 @@ def main():
     camera2 = sens.ChCameraSensor(
                 patch.GetGroundBody(),  # body camera is attached to
                 30,                      # update rate in Hz
-                chrono.ChFrameD(chrono.ChVectorD(-3.09,-2.89,1.28), 
-                chrono.Q_from_Euler123(chrono.ChVectorD(.0,.45,chrono.CH_C_PI_4+.1))),  # offset pose   ------ not sure what CH_C_PI_2 is
-                1920,                                                           # image width
-                1080,                                                            # image height
-                1.2,                                                       # FOV
+                chrono.ChFrameD(chrono.ChVectorD(init_loc_x + 2, init_loc_y, 10), 
+                chrono.Q_from_AngAxis(chrono.CH_C_PI_2, chrono.ChVectorD(0, 1, 0))),  # offset pose   ------ not sure what CH_C_PI_2 is
+                1280,                                                           # image width
+                720,                                                            # image height
+                3.14 / 4,                                                       # FOV
                 2)                                                              # super sample diameter
     camera2.SetName("Camera Sensor 2")
-    if vis:
-        camera2.PushFilter(sens.ChFilterVisualize(640, 360))
+    # camera2.PushFilter(sens.ChFilterVisualize(1280, 720))
     if save_sensor_data:
         camera2.PushFilter(sens.ChFilterSave(sensor_data_dir + "cam2/"))
     manager.AddSensor(camera2)
 
 
     noise_model = sens.ChNoiseNone()
-    gps_noise_model = sens.ChNoiseNone()
-    #gps_noise_model = sens.ChNoiseNormal(chrono.ChVectorD(0, 0, 0), chrono.ChVectorD(0.6, 0.6, 0.6))
     imu_offset_pose = chrono.ChFrameD(chrono.ChVectorD(0, 0, 0), chrono.Q_from_AngAxis(0, chrono.ChVectorD(1, 0, 0)))
     gps_offset_pose = chrono.ChFrameD(chrono.ChVectorD(0, 0, 0), chrono.Q_from_AngAxis(0, chrono.ChVectorD(1, 0, 0)))
     gps_reference = chrono.ChVectorD(-89.400, 43.070, 260.0)
@@ -431,16 +434,10 @@ def main():
     manager.AddSensor(mag)
 
     # gps
-    gps = sens.ChGPSSensor(vehicle.GetChassisBody(), 10, gps_offset_pose, gps_reference, gps_noise_model)
+    gps = sens.ChGPSSensor(vehicle.GetChassisBody(), 10, gps_offset_pose, gps_reference, noise_model)
     gps.SetName("GPS")
     gps.PushFilter(sens.ChFilterGPSAccess())
     manager.AddSensor(gps)
-
-    #Ground Truth
-    groundTruth = sens.ChGPSSensor(vehicle.GetChassisBody(), 10, gps_offset_pose, gps_reference, gps_noise_model)
-    groundTruth.SetName("groundTruth")
-    groundTruth.PushFilter(sens.ChFilterGPSAccess())
-    manager.AddSensor(groundTruth)
 
     # label the cones that were added to the scene with semantic information
     manager.ReconstructScenes()
@@ -470,9 +467,6 @@ def main():
     gps_generator = ChGPSSensor_DataGeneratorFunctor("~/output/gps/data", gps)
     driver.AddDataGenerator(gps_generator, 10)
 
-    groundTruth_generator = ChGPSSensor_DataGeneratorFunctor("~/output/groundTruth/data", groundTruth)
-    driver.AddDataGenerator(groundTruth_generator, 10)
-
     inputs_parser = ChDriverInputs_DataParserFunctor(driver)
     driver.AddDataParser(inputs_parser)
 
@@ -488,18 +482,6 @@ def main():
 
     # Simulation loop
     realtime_timer = chrono.ChRealtimeStepTimer()
-
-    psteering = 0
-    pthrottle = 0
-    pbraking = 0
-
-    throttle_time = 1.0
-    steering_time = 1.0
-    braking_time = .5
-    delta_throttle = 1.0 * step_size / throttle_time
-    delta_braking = 1.0 * step_size / braking_time
-    delta_steering = 2.0 * step_size / steering_time
-
     while True:
         time = vehicle.GetSystem().GetChTime()
     
@@ -515,17 +497,6 @@ def main():
 
         # Get driver inputs
         driver_inputs = driver.GetInputs()
-
-        #clamp driver inputs
-        driver_inputs.m_steering
-
-        pthrottle = np.clip(driver_inputs.m_throttle,pthrottle-delta_throttle,pthrottle+delta_throttle)
-        psteering = np.clip(driver_inputs.m_steering,psteering-delta_steering,psteering+delta_steering)
-        pbraking = np.clip(driver_inputs.m_braking,pbraking-delta_braking,pbraking+delta_braking)
-
-        driver_inputs.m_throttle = pthrottle
-        driver_inputs.m_steering = psteering
-        driver_inputs.m_braking = pbraking
 
         # Update modules (process inputs from other modules)
         driver.Synchronize(time)
@@ -554,18 +525,17 @@ def main():
 # The path to the Chrono data directory containing various assets (meshes, textures, data files)
 # is automatically set, relative to the default location of this demo.
 # If running from a different directory, you must change the path to the data directory with: 
-data_folder = os.path.join(os.path.dirname(os.path.abspath(__file__)), "data", "")
+data_folder = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "data", "")
 chrono.SetChronoDataPath(data_folder)
 veh.SetDataPath(chrono.GetChronoDataPath() + 'vehicle/')
 
-vis = True
-save_sensor_data = True
+create_semantic_maps = False
+save_sensor_data = False
 sensor_data_dir = "sensor_output/"
 
 num_cones = 100
-cones_from_file = False
-# cone_path_file = "autonomy-toolkit/paths/cone_path_01.csv"
-cone_path_file = "autonomy-toolkit/paths/cone_path_iros.csv"
+cones_from_file = True
+cone_path_file = "data/paths/cone_paths_0.csv"
 
 cone_offset_x = 0
 cone_offset_y = 0
@@ -585,10 +555,12 @@ steering = 0.0
 red_cone_assets = list()
 green_cone_assets = list()
 
-cones = list()
+
+red_cones = list()
+green_cones = list()
 
 # Initial vehicle location
-init_loc_x = -2.2
+init_loc_x = -2.5
 init_loc_y = 0.5
 init_angle_z = 1.57
